old_root_files/requirements.txt
MetaTrader5
pandas
python-dotenv
numpy
yfinance
